# internal/controller/restapi/v1/contact.go
// Package v1 provides contact REST API handlers.
package v1

import (
	"errors"
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
	"github.com/nabilfikrisp/go-crud/internal/controller/restapi/v1/request"
	"github.com/nabilfikrisp/go-crud/internal/dto"
	"github.com/nabilfikrisp/go-crud/internal/entity"
)

// @Summary      Create a new contact
// @Description  Create a contact with the provided details
// @Tags         contacts
// @Accept       json
// @Produce      json
// @Param        request body request.CreateContact true "Create Contact Body"
// @Success      201  {object}  entity.Contact
// @Failure      400  {object}  response.Error  "Invalid request or validation error"
// @Failure      500  {object}  response.Error  "Internal server error"
// @Router       /contacts [post]
func (r *V1) createContact(c *gin.Context) {
	var body request.CreateContact

	if err := c.ShouldBindJSON(&body); err != nil {
		errorResponse(c, http.StatusBadRequest, "Invalid request body: "+err.Error())
		return
	}

	if err := r.v.Struct(body); err != nil {
		r.l.Error(err, "restapi - v1 - createContact")
		validationErrorResponse(c, err)
		return
	}

	contact, err := r.c.Create(c.Request.Context(), dto.ContactCreate{
		FirstName:    body.FirstName,
		LastName:     body.LastName,
		Email:        body.Email,
		PhoneNumber:  body.PhoneNumber,
		Relationship: body.Relationship,
	})
	if err != nil {
		r.l.Error(err, "restapi - v1 - createContact")
		errorResponse(c, http.StatusInternalServerError, "Failed to create contact")
		return
	}

	c.JSON(http.StatusCreated, contact)
}

// @Summary      Get a contact by ID
// @Description  Retrieve details of a single contact using its unique ID
// @Tags         contacts
// @Produce      json
// @Param        id   path      string  true  "Contact ID"
// @Success      200  {object}  entity.Contact
// @Failure      400  {object}  response.Error  "Invalid relationship"
// @Failure      404  {object}  response.Error  "Contact not found"
// @Failure      500  {object}  response.Error  "Internal server error"
// @Router       /contacts/{id} [get]
func (r *V1) getContactByID(c *gin.Context) {
	id := c.Param("id")

	contact, err := r.c.GetByID(c.Request.Context(), id)
	if err != nil {
		if errors.Is(err, entity.ErrContactRelationshipInvalid) {
			errorResponse(c, http.StatusBadRequest, "Invalid relationship")
			return
		}

		if errors.Is(err, entity.ErrContactNotFound) {
			errorResponse(c, http.StatusNotFound, "Contact not found")
			return
		}

		r.l.Error(err, "restapi - v1 - getContactByID")
		errorResponse(c, http.StatusInternalServerError, "Failed to get contact")
		return
	}

	c.JSON(http.StatusOK, contact)
}

// @Summary      List contacts
// @Description  Retrieve a paginated list of contacts with optional filtering
// @Tags         contacts
// @Produce      json
// @Param        first_name    query     string  false  "Filter by first name"
// @Param        last_name     query     string  false  "Filter by last name"
// @Param        email         query     string  false  "Filter by email"
// @Param        phone_number  query     string  false  "Filter by phone number"
// @Param        relationship  query     string  false  "Filter by relationship (friend, family, work, etc)"
// @Param        limit         query     int     false  "Page limit (default 10)"
// @Param        offset        query     int     false  "Page offset (default 0)"
// @Success      200  {object}  map[string]interface{} "Returns {contacts: []entity.Contact, total: int}"
// @Failure      400  {object}  response.Error
// @Failure      500  {object}  response.Error
// @Router       /contacts [get]
func (r *V1) listContacts(c *gin.Context) {
	filter := request.ContactFilter{
		Limit:  new(uint64(10)),
		Offset: new(uint64(0)),
	}

	if v := c.Query("first_name"); v != "" {
		filter.FirstName = new(v)
	}

	if v := c.Query("last_name"); v != "" {
		filter.LastName = new(v)
	}

	if v := c.Query("email"); v != "" {
		filter.Email = new(v)
	}

	if v := c.Query("phone_number"); v != "" {
		filter.PhoneNumber = new(v)
	}

	if v := c.Query("relationship"); v != "" {
		rel := entity.ContactRelationship(v)
		if !rel.Valid() {
			errorResponse(c, http.StatusBadRequest, "invalid relationship")
			return
		}
		filter.Relationship = &rel
	}

	if v, err := strconv.ParseUint(c.Query("limit"), 10, 64); err == nil {
		filter.Limit = &v
	}
	if v, err := strconv.ParseUint(c.Query("offset"), 10, 64); err == nil {
		filter.Offset = &v
	}

	contacts, total, err := r.c.List(c.Request.Context(), filter.ToDTO())
	if err != nil {
		r.l.Error(err, "restapi - v1 - listContacts")
		errorResponse(c, http.StatusInternalServerError, "failed to list contacts")
		return
	}
	c.JSON(http.StatusOK, gin.H{
		"contacts": contacts,
		"total":    total,
	})
}

// @Summary      Update a contact
// @Description  Update existing contact details by ID
// @Tags         contacts
// @Accept       json
// @Produce      json
// @Param        id      path      string                true  "Contact ID"
// @Param        request body      request.UpdateContact true  "Update Contact Body"
// @Success      200  {object}  entity.Contact
// @Failure      400  {object}  response.Error
// @Failure      404  {object}  response.Error
// @Failure      500  {object}  response.Error
// @Router       /contacts/{id} [put]
func (r *V1) updateContact(c *gin.Context) {
	id := c.Param("id")

	var body request.UpdateContact

	if err := c.ShouldBindJSON(&body); err != nil {
		errorResponse(c, http.StatusBadRequest, "Invalid request body: "+err.Error())
		return
	}

	if err := r.v.Struct(body); err != nil {
		r.l.Error(err, "restapi - v1 - updateContact")
		validationErrorResponse(c, err)
		return
	}

	contact, err := r.c.Update(c.Request.Context(), id, dto.ContactUpdate{
		FirstName:    body.FirstName,
		LastName:     body.LastName,
		Email:        body.Email,
		PhoneNumber:  body.PhoneNumber,
		Relationship: body.Relationship,
	})
	if err != nil {
		if errors.Is(err, entity.ErrContactNotFound) {
			errorResponse(c, http.StatusNotFound, "Contact not found")
			return
		}

		if errors.Is(err, entity.ErrContactRelationshipInvalid) {
			errorResponse(c, http.StatusBadRequest, "Invalid relationship")
			return
		}

		r.l.Error(err, "restapi - v1 - updateContact")
		errorResponse(c, http.StatusInternalServerError, "Failed to update contact")
		return
	}

	c.JSON(http.StatusOK, contact)
}

// @Summary      Delete a contact
// @Description  Remove a contact from the system by ID
// @Tags         contacts
// @Param        id   path      string  true  "Contact ID"
// @Success      204  "No Content"
// @Failure      404  {object}  response.Error
// @Failure      500  {object}  response.Error
// @Router       /contacts/{id} [delete]
func (r *V1) deleteContact(c *gin.Context) {
	id := c.Param("id")

	if err := r.c.Delete(c.Request.Context(), id); err != nil {
		if errors.Is(err, entity.ErrContactNotFound) {
			errorResponse(c, http.StatusNotFound, "Contact not found")
			return
		}

		r.l.Error(err, "restapi - v1 - deleteContact")
		errorResponse(c, http.StatusInternalServerError, "Failed to delete contact")
		return
	}

	c.Status(http.StatusNoContent)
}